+++ platform-tools-sdk/cargo-build-sbf/src/main.rs
@@ -44,6 +44,7 @@ struct Config<'a> {
     workspace: bool,
     jobs: Option<String>,
     arch: &'a str,
+    optimize_size: bool,
 }
 
 impl Default for Config<'_> {
@@ -74,6 +75,7 @@ impl Default for Config<'_> {
             workspace: false,
             jobs: None,
             arch: "v0",
+            optimize_size: false,
         }
     }
 }
@@ -782,6 +784,9 @@ fn build_solana_package(
     if config.remap_cwd && !config.debug {
         target_rustflags = Cow::Owned(format!("{} -Zremap-cwd-prefix=", &target_rustflags));
     }
+    if config.optimize_size {
+        target_rustflags = Cow::Owned(format!("{} -C opt-level=s", &target_rustflags));
+    }
     if config.debug {
         // Replace with -Zsplit-debuginfo=packed when stabilized.
         target_rustflags = Cow::Owned(format!("{} -g", &target_rustflags));
@@ -1172,6 +1177,12 @@ fn main() {
                 .default_value("v0")
                 .help("Build for the given target architecture"),
         )
+        .arg(
+            Arg::new("optimize_size")
+                .long("optimize-size")
+                .takes_value(false)
+                .help("Optimize program for size. This option may reduce program size, potentially increasing CU consumption.")
+        )
         .get_matches_from(args);
 
     let sbf_sdk: PathBuf = matches.value_of_t_or_exit("sbf_sdk");
@@ -1241,6 +1252,7 @@ fn main() {
         workspace: matches.is_present("workspace"),
         jobs: matches.value_of_t("jobs").ok(),
         arch: matches.value_of("arch").unwrap(),
+        optimize_size: matches.is_present("optimize_size"),
     };
     let manifest_path: Option<PathBuf> = matches.value_of_t("manifest_path").ok();
     if config.verbose {